Updated: scouts and news controller, interfaces and repositories Added: scoutspal configurantion helper and scout events repository and interface

+++ ScoutsPal.Services.EventsManagerAPI/Services/NewsRepository.cs
@@ -203,7 +203,7 @@ namespace ScoutsPal.Services.EventsManagerAPI.Services
         /// </summary>
         /// <param name="groupId"></param>
         /// <returns></returns>
-        public IEnumerable<News> GetNewsByGroup(int groupId)
+        public IEnumerable<News> GetNewsByGroup(long groupId)
         {
             List<News> newsList = new List<News>();
